+++ userscript/src/typings.ts
@@ -5,3 +5,8 @@ export const XMLHttpRequest = window.XMLHttpRequest as any as typeof window.XMLH
     _open: typeof window.XMLHttpRequest["prototype"]["open"];
   };
 };
+
+export enum SocketTypes {
+  online,
+  party,
+}