+++ userscript/src/utils.ts
@@ -1,3 +1,5 @@
+import type { Socket } from "socket.io-client";
+
 /* Small funtion to test if you are in a game. */
 export const inGame = () =>
   app.matchStarted && app.client.socket && app.client.socket.readyState == WebSocket.OPEN;
@@ -7,4 +9,8 @@ export function setHash(id: string, name: string, pass: string) {
   window.location.hash = pass
     ? `${id}&${encodeURIComponent(name)}&${encodeURIComponent(pass)}`
     : `${id}&${encodeURIComponent(name)}`;
+}
+
+export function io(url: string) {
+  return (window as any).io(url) as Socket;
 }